버그 수정 및 FPS 최적화 작업

+++ API_Project/structs.h
@@ -4,8 +4,8 @@
 using namespace std;
 struct Vector2D
 {
-	float x = 0.0f;
-	float y = 0.0f;
+	double x = 0.0f;
+	double y = 0.0f;
 };
 
 struct Animation
